requirements.txt
Pillow>=9.0.0
PyMuPDF>=1.20.0
python-docx>=0.8.11
openpyxl>=3.0.0
